Update for new game version 1.0.70024, added options to reverse changes to collateral damage calculation and competitive mode character yields.

+++ BetterAI/BetterAIClientUI.cs
@@ -461,7 +461,7 @@ namespace BetterAI
 
                                         for (VegetationType eLoopVegetation = 0; (eLoopVegetation < Infos.vegetationNum() && !(bAnyValid)); eLoopVegetation++)
                                         {
-                                            bAnyValid = bAnyValid || improvement.mabVegetationValid[(int)eLoopVegetation];
+                                            bAnyValid = bAnyValid || improvement.mabVegetationValid[eLoopVegetation];
                                         }
 
                                         if (!(bAnyValid))
